Tests/LibWeb/Text/input/HTML/invalid-url-attribute-values.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    test(() => {
        const elementList = [
            { "img": "currentSrc" },
            { "object": "data" },
        ];
        for (const elementDescriptor of elementList) {
            [elementName, propertyName] = Object.entries(elementDescriptor)[0];
            const element = document.createElement(elementName);
            println(`${elementName}.${propertyName} default value: '${element[propertyName]}'`);
        }
    });
</script>
