Tests/LibWeb/Text/expected/HTML/invalid-url-attribute-values.txt

img.currentSrc default value: ''
object.data default value: ''
